Primary Cox PH regression with multiple starting points and optional combinations of null, stratification, competing risks, multiplicative log-linear model, and no derivative calculation. LogLik_Cox_PH_Omnibus Performs the calls to calculation functions, Structures the Cox PH regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives

Primary Cox PH regression with multiple starting points and optional combinations of null, stratification, competing risks, multiplicative log-linear model, and no derivative calculation. LogLik_Cox_PH_Omnibus Performs the calls to calculation functions, Structures the Cox PH regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives

Usage

LogLik_Cox_PH_Omnibus(
  Term_n,
  tform,
  a_ns,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  lr,
  maxiters,
  guesses,
  halfmax,
  epsilon,
  dbeta_cap,
  abs_max,
  dose_abs_max,
  deriv_epsilon,
  df_groups,
  tu,
  double_step,
  change_all,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  ties_method,
  nthreads,
  STRATA_vals,
  cens_weight,
  cens_thres,
  strata_bool,
  basic_bool,
  null_bool,
  CR_bool,
  single_bool,
  constraint_bool,
  gmix_theta,
  gmix_term,
  Lin_Sys,
  Lin_Res
)

Arguments

Term_n

Term numbers

tform

subterm types

a_ns

matrix of starting values

x_all

covariate matrix

dfc

vector matching subterm number to matrix column

fir

st term number

der_iden

subterm number for derivative tests

modelform

string model identifier

lr

learning rate for newton step toward 0 derivative

maxiters

list of maximum number of iterations

guesses

number of initial

halfmax

maximum number of half steps

epsilon

minimum acceptable maximum parameter change

dbeta_cap

learning rate for newton step toward 0 log-likelihood

abs_max

Maximum allowed parameter change

dose_abs_max

Maximum allowed threshold parameter change

deriv_epsilon

threshold for near-zero derivative

df_groups

matrix with time and event information

tu

Event time vector

double_step

controls the step calculation, 0 for independent changes, 1 for solving b=Ax with complete matrices

change_all

boolean if every parameter is being updated

verbose

boolean for additional printing

debugging

additional boolean for verbosity in testing

KeepConstant

binary vector to identify constant parameters

term_tot

total number of terms

ties_method

Ties method

nthreads

number of threads available

STRATA_vals

vector of strata identifier values

cens_weight

vector of censoring weights

cens_thres

threshold to add competing event to risk group

strata_bool

boolean for stratification

basic_bool

boolean for multiplicative log-linear model

null_bool

boolean for null model

CR_bool

boolean for competing risks

single_bool

boolean for single calculation without derivatives

constraint_bool

boolean for system of linear equality constraints used

gmix_theta

theta value for geometric-mixture model

gmix_term

list of 0/1 to identify which terms to set as Relative Risk (0) or Excess Risk (1)

Lin_Sys

Matrix containing coefficients for system of linear constraints

Lin_Res

Vector containing constants for system of linear constraints

Value

List of final results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information