Saves information about a run to a log file

Model_Results_Log saves information about the data, results, model, computer, software, and date to an external file. Intended to make reproduction of results easier

Usage

Model_Results_Log(
  log_file = "out.log",
  df = data.table(),
  out_list = list(),
  noprint = TRUE
)

Arguments

log_file

file to save log to

df

a data.table containing the columns of interest

out_list

list output from a regression, used to build results table and pull out convergence values

noprint

boolean, if true the file is saved to the log file, if false the output is printed to the console INSTEAD of being saved.

Value

null, prints to screen or saves to file

See also