Performs basic Cox Proportional Hazards regression with a multiplicative log-linear model

RunCoxRegression_Basic uses user provided data, time/event columns, vectors specifying the model, and options to control the convergence and starting positions

Usage

RunCoxRegression_Basic(
  df,
  time1,
  time2,
  event0,
  names,
  keep_constant,
  a_n,
  der_iden,
  control
)

Arguments

df

a data.table containing the columns of interest

time1

column used for time period starts

time2

column used for time period end

event0

column used for event status

names

columns for elements of the model, used to identify data columns

keep_constant

binary values to denote which parameters to change

a_n

list of initial parameter values, used to determine number of parameters

der_iden

number for the subterm to test derivative at, only used for testing runs with a single varying parameter, should be smaller than total number of parameters

control

list of parameters controlling the convergence, see Def_Control() for options or vignette("starting_description")

Value

returns a list of the final results

See also

Examples

library(data.table)
## basic example code reproduced from the starting-description vignette

df <- data.table::data.table("UserID"=c(112, 114, 213, 214, 115, 116, 117),
           "Starting_Age"=c(18,  20,  18,  19,  21,  20,  18),
             "Ending_Age"=c(30,  45,  57,  47,  36,  60,  55),
          "Cancer_Status"=c(0,   0,   1,   0,   1,   0,   0),
                      "a"=c(0,   1,   1,   0,   1,   0,   1),
                      "b"=c(1,   1.1, 2.1, 2,   0.1, 1,   0.2),
                      "c"=c(10,  11,  10,  11,  12,  9,   11),
                      "d"=c(0,   0,   0,   1,   1,   1,   1))
# For the interval case
time1 <- "Starting_Age"
time2 <- "Ending_Age"
event <- "Cancer_Status"
names <- c('a','b','c','d')
a_n <- c(1.1, -0.1, 0.2, 0.5) #used to test at a specific point

keep_constant <- c(0,0,0,0)
der_iden <- 0

control <- list("ncores"=2,'lr' = 0.75,'maxiter' = 5,'halfmax' = 5,
   'epsilon' = 1e-3,'dbeta_max' = 0.5,'deriv_epsilon' = 1e-3, 'abs_max'=1.0,
   'change_all'=TRUE,'dose_abs_max'=100.0,'verbose'=FALSE,
   'ties'='breslow','double_step'=1)

e <- RunCoxRegression_Basic(df, time1, time2, event, names, keep_constant,
                            a_n, der_iden, control)